Server: Simple Unit Test
