+++ server/ToysRevive-server.cabal
@@ -26,6 +26,7 @@ source-repository head
 library
   exposed-modules:
       Server
+      API.ServerMessage
   other-modules:
       Paths_ToysRevive_server
   autogen-modules:
@@ -71,6 +72,7 @@ test-suite ToysRevive-server-test
   type: exitcode-stdio-1.0
   main-is: Spec.hs
   other-modules:
+      TestServer
       Paths_ToysRevive_server
   autogen-modules:
       Paths_ToysRevive_server
@@ -81,11 +83,14 @@ test-suite ToysRevive-server-test
       DeriveGeneric
   ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints -threaded -rtsopts -with-rtsopts=-N
   build-depends:
-      ToysRevive-server
+      HUnit
+    , ToysRevive-server
     , aeson
     , base >=4.7 && <5
     , servant
     , servant-server
+    , test-framework
+    , test-framework-hunit
     , wai
     , warp
   default-language: Haskell2010
